implemented reviewer comments
