+++ src/app/static/src/app/components/plots/utils.ts
@@ -4,6 +4,8 @@ import {Dict, Filter, NumericRange} from "../../types";
 import numeral from 'numeral';
 import i18next from "i18next";
 import {Language} from "../../store/translations/locales";
+import {BarchartSelections} from "../../store/plottingSelections/plottingSelections";
+import {flattenOptionsIdsByHierarchy} from "../../utils"
 
 export const getColor = (value: number,
                          metadata: ChoroplethIndicatorMetadata,
@@ -216,6 +218,35 @@ export const formatLegend = function (text: string | number, format: string, sca
         } else text = text.toString()
     }
     return text
+}
+
+export const flattenXAxisFilterOptionIds = (selections: BarchartSelections, filters: Filter[]) => {
+    const xAxisId = selections?.xAxisId
+    let ids: string[] = []
+    if (xAxisId && filters?.length) {
+        const filter = filters.find((f: Filter) => f.id === xAxisId)
+        if (filter?.options.length) {
+            ids = flattenOptionsIdsByHierarchy(filter.options)
+        }
+    }
+    return ids
+}
+
+
+export const updateSelectionsAndXAxisOrder = (data: BarchartSelections, selections: BarchartSelections, flattenedXAxisFilterOptionIds: string[], updateSelections: (data: {payload: BarchartSelections}) => void) => {
+    const payload = {...selections, ...data}
+    if (data.xAxisId && data.selectedFilterOptions) {
+        const {xAxisId, selectedFilterOptions} = data
+        if (selectedFilterOptions[xAxisId] && flattenedXAxisFilterOptionIds.length) {
+            // Sort the selected filter values according to the order given the barchart filters
+            const updatedFilterOptions = [...selectedFilterOptions[xAxisId]].sort((a: FilterOption, b: FilterOption) => {
+                return flattenedXAxisFilterOptionIds.indexOf(a.id) - flattenedXAxisFilterOptionIds.indexOf(b.id);
+            });
+            payload.selectedFilterOptions[xAxisId] = updatedFilterOptions
+        }
+    }
+    // if unable to do the above, just updates the barchart as normal
+    updateSelections({payload})
 }
 
 export const filterConfig = (currentLanguage: Language, filters: Filter[]) => {
